Using Contract to clean code, clean SearchAdapter
