+++ app/src/main/AndroidManifest.xml
@@ -1,11 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
-    <uses-permission android:name="android.permission.INTERNET"/>
+
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:allowBackup="true"
         android:name=".MyApp"
+        android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
@@ -15,27 +16,24 @@
         android:theme="@style/Theme.MovieDB">
         <activity
             android:name=".screen.login.LoginActivity"
-            android:exported="false">
-        </activity>
+            android:exported="false"/>
         <activity
             android:name=".screen.splash.SplashActivity"
-            android:exported="true">
-
-        </activity>
-        <activity
-            android:name=".screen.MainActivity"
-            android:exported="true">
-
-        </activity>
-
-        <activity android:name=".screen.watchMovie.WatchMovieActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+
         </activity>
+        <activity
+            android:name=".screen.MainActivity"
+            android:exported="true"/>
+
+        <activity
+            android:name=".screen.watchMovie.WatchMovieActivity"
+            android:exported="true"/>
 
     </application>
