--- app/src/main/java/com/sun/moviedb/screen/home/HomeView.kt
@@ -1,11 +1,0 @@
-package com.sun.moviedb.screen.home
-
-import com.sun.moviedb.data.model.Item
-import com.sun.moviedb.utils.base.BaseView
-
-interface HomeView : BaseView {
-    fun showNewestMoviesLoading(isLoading: Boolean)
-    fun showSeriesMoviesLoading(isLoading: Boolean)
-    fun showNewestMovies(items: List<Item>)
-    fun showSeriesMovies(items: List<Item>, currentPage: Int, totalPage: Int)
-}
